backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-multipart==0.0.20
GitPython==3.1.44
radon==6.0.1
bandit==1.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.5
httpx==0.28.1
pytest==8.3.4
ruff==0.8.4
